requirements.txt
# Smart AI Study Assistant - Dependencies
# Install with: pip install -r requirements.txt

streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
